+++ test/tenso_test.js
@@ -1,20 +1,22 @@
 var expect = require('chai').expect,
     hippie = require( "hippie" ),
     tenso  = require( "../lib/tenso" ),
-    routes = require( "./routes.js" ),
-    app    = tenso( {routes: routes, logs: {level: "error"}} );
+    routes = require( "./routes.js" );
 
-function api () {
-	return hippie()
-		.json()
-		.base("http://localhost:8000")
-		.expectHeader("Content-Type", "application/json");
+function api ( port, not_json ) {
+	var obj = hippie().base("http://localhost:" + port)
+
+	return not_json ? obj : obj.expectHeader("Content-Type", "application/json").json();
 }
 
 describe("Permissions", function () {
+	var port = 8001;
+
+	tenso( {port: port, routes: routes, logs: {level: "error"}} );
+
 	describe("GET /", function () {
 		it("returns an array of endpoints", function (done) {
-			api()
+			api( port )
 				.get("/")
 				.expectStatus(200)
 				.expectValue("data.link", [])
@@ -30,7 +32,7 @@ describe("Permissions", function () {
 
 	describe("GET /invalid", function () {
 		it("returns a 'not found' error", function (done) {
-			api()
+			api( port )
 				.get("/invalid")
 				.expectStatus(404)
 				.expectValue("data", null)
@@ -45,7 +47,7 @@ describe("Permissions", function () {
 
 	describe("DELETE /", function () {
 		it("returns a 'method not allowed' error", function (done) {
-			api()
+			api( port )
 				.del("/")
 				.expectStatus(405)
 				.expectValue("data", null)
@@ -60,7 +62,7 @@ describe("Permissions", function () {
 
 	describe("POST /", function () {
 		it("returns a 'method not allowed' error", function (done) {
-			api()
+			api( port )
 				.post("/")
 				.expectStatus(405)
 				.expectValue("data", null)
@@ -75,7 +77,7 @@ describe("Permissions", function () {
 
 	describe("PUT /", function () {
 		it("returns a 'method not allowed' error", function (done) {
-			api()
+			api( port )
 				.put("/")
 				.expectStatus(405)
 				.expectValue("data", null)
@@ -90,7 +92,7 @@ describe("Permissions", function () {
 
 	describe("PATCH /", function () {
 		it("returns a 'method not allowed' error", function (done) {
-			api()
+			api( port )
 				.patch("/")
 				.expectStatus(405)
 				.expectValue("data", null)
@@ -105,9 +107,13 @@ describe("Permissions", function () {
 });
 
 describe("Pagination", function () {
+	var port = 8002;
+
+	tenso( {port: port, routes: routes, logs: {level: "error"}} );
+
 	describe("GET /empty", function () {
 		it("returns an empty array", function (done) {
-			api()
+			api( port )
 				.get("/empty")
 				.expectStatus(200)
 				.expectValue("data.link", [])
@@ -123,10 +129,10 @@ describe("Pagination", function () {
 
 	describe("GET /items", function () {
 		it("returns page 1/3 of an array of numbers", function (done) {
-			api()
+			api( port )
 				.get("/items")
 				.expectStatus(200)
-				.expectValue("data.link", [{ uri: "http://localhost:8000/items?page=2&page_size=5", rel: "next" }, { uri: "http://localhost:8000/items?page=3&page_size=5", rel: "last" }])
+				.expectValue("data.link", [{ uri: "http://localhost:" + port + "/items?page=2&page_size=5", rel: "next" }, { uri: "http://localhost:" + port + "/items?page=3&page_size=5", rel: "last" }])
 				.expectValue("data.result", [1,2,3,4,5])
 				.expectValue("error", null)
 				.expectValue("status", 200)
@@ -139,10 +145,10 @@ describe("Pagination", function () {
 
 	describe("GET /items?page=2&page_size=5", function () {
 		it("returns page 2/3 of an array of numbers", function (done) {
-			api()
+			api( port )
 				.get("/items?page=2&page_size=5")
 				.expectStatus(200)
-				.expectValue("data.link", [{ uri: "http://localhost:8000/items?page=1&page_size=5", rel: "first" }, { uri: "http://localhost:8000/items?page=3&page_size=5", rel: "last" }])
+				.expectValue("data.link", [{ uri: "http://localhost:" + port + "/items?page=1&page_size=5", rel: "first" }, { uri: "http://localhost:" + port + "/items?page=3&page_size=5", rel: "last" }])
 				.expectValue("data.result", [6,7,8,9,10])
 				.expectValue("error", null)
 				.expectValue("status", 200)
@@ -155,10 +161,10 @@ describe("Pagination", function () {
 
 	describe("GET /items?page=3&page_size=5", function () {
 		it("returns page 3/3 of an array of numbers", function (done) {
-			api()
+			api( port )
 				.get("/items?page=3&page_size=5")
 				.expectStatus(200)
-				.expectValue("data.link", [{ uri: "http://localhost:8000/items?page=1&page_size=5", rel: "first" }, { uri: "http://localhost:8000/items?page=2&page_size=5", rel: "prev" }])
+				.expectValue("data.link", [{ uri: "http://localhost:" + port + "/items?page=1&page_size=5", rel: "first" }, { uri: "http://localhost:" + port + "/items?page=2&page_size=5", rel: "prev" }])
 				.expectValue("data.result", [11,12,13,14,15])
 				.expectValue("error", null)
 				.expectValue("status", 200)
@@ -171,10 +177,10 @@ describe("Pagination", function () {
 
 	describe("GET /items?page=4&page_size=5", function () {
 		it("returns page 4/3 of an array of numbers (empty)", function (done) {
-			api()
+			api( port )
 				.get("/items?page=4&page_size=5")
 				.expectStatus(200)
-				.expectValue("data.link", [{ uri: "http://localhost:8000/items?page=1&page_size=5", rel: "first" }, { uri: "http://localhost:8000/items?page=3&page_size=5", rel: "last" }])
+				.expectValue("data.link", [{ uri: "http://localhost:" + port + "/items?page=1&page_size=5", rel: "first" }, { uri: "http://localhost:" + port + "/items?page=3&page_size=5", rel: "last" }])
 				.expectValue("data.result", [])
 				.expectValue("error", null)
 				.expectValue("status", 200)
@@ -187,10 +193,10 @@ describe("Pagination", function () {
 
 	describe("GET /items?email=user@domain.com", function () {
 		it("returns page 1/3 of an array of numbers, preserving the query string via encoding", function (done) {
-			api()
+			api( port )
 				.get("/items?email=user@domain.com")
 				.expectStatus(200)
-				.expectValue("data.link", [{ uri: "http://localhost:8000/items?email=user%40domain.com&page=2&page_size=5", rel: "next" }, { uri: "http://localhost:8000/items?email=user%40domain.com&page=3&page_size=5", rel: "last" }])
+				.expectValue("data.link", [{ uri: "http://localhost:" + port + "/items?email=user%40domain.com&page=2&page_size=5", rel: "next" }, { uri: "http://localhost:" + port + "/items?email=user%40domain.com&page=3&page_size=5", rel: "last" }])
 				.expectValue("data.result", [1,2,3,4,5])
 				.expectValue("error", null)
 				.expectValue("status", 200)
@@ -203,12 +209,16 @@ describe("Pagination", function () {
 });
 
 describe("Hypermedia", function () {
+	var port = 8003;
+
+	tenso( {port: port, routes: routes, logs: {level: "error"}} );
+
 	describe("GET /something", function () {
 		it("returns an entity that has hypermedia properties", function (done) {
-			api()
+			api( port )
 				.get("/something")
 				.expectStatus(200)
-				.expectValue("data.link", [{ uri: "http://localhost:8000/users/123", rel: "related" }, { uri: "http://source.tld", rel: "related" }])
+				.expectValue("data.link", [{ uri: "http://localhost:" + port + "/users/123", rel: "related" }, { uri: "http://source.tld", rel: "related" }])
 				.expectValue("data.result", {"title": "This is a title", "body": "Where is my body?"})
 				.expectValue("error", null)
 				.expectValue("status", 200)
@@ -218,4 +228,74 @@ describe("Hypermedia", function () {
 				});
 		});
 	});
+});
+
+describe("Basic Auth", function () {
+	var port = 8004;
+
+	tenso( {port: port, routes: routes, logs: {level: "error"}, auth: {basic: {enabled: true, list:["test:123"]}}} );
+
+	describe( "GET /", function () {
+		it( "returns an array of endpoints (authorized)", function ( done ) {
+			api( port )
+				.auth('test', '123')
+				.get( "/" )
+				.expectStatus( 200 )
+				.expectValue( "data.link", [] )
+				.expectValue( "data.result", ["/items"] )
+				.expectValue( "error", null )
+				.expectValue( "status", 200 )
+				.end( function ( err ) {
+					if ( err ) throw err;
+					done();
+				} );
+		} );
+	} );
+
+	describe( "GET /", function () {
+		it( "returns an 'unauthorized' error", function ( done ) {
+			api( port, true )
+				.get( "/" )
+				.expectStatus( 401 )
+				.end( function ( err ) {
+					if ( err ) throw err;
+					done();
+				} );
+		} );
+	} );
+});
+
+describe("OAuth2 Token Bearer", function () {
+	var port = 8005;
+
+	tenso( {port: port, routes: routes, logs: {level: "error"}, auth: {bearer: {enabled: true, tokens:["abc-123"]}}} );
+
+	describe( "GET /", function () {
+		it( "returns an array of endpoints (authorized)", function ( done ) {
+			api( port )
+				.header('Authorization', 'Bearer abc-123')
+				.get( "/" )
+				.expectStatus( 200 )
+				.expectValue( "data.link", [] )
+				.expectValue( "data.result", ["/items"] )
+				.expectValue( "error", null )
+				.expectValue( "status", 200 )
+				.end( function ( err ) {
+					if ( err ) throw err;
+					done();
+				} );
+		} );
+	} );
+
+	describe( "GET /", function () {
+		it( "returns an 'unauthorized' error", function ( done ) {
+			api( port, true )
+				.get( "/" )
+				.expectStatus( 401 )
+				.end( function ( err ) {
+					if ( err ) throw err;
+					done();
+				} );
+		} );
+	} );
 });